Split sync into push and pull

+++ .bashrc
@@ -86,8 +86,9 @@ export DOTNET_CLI_TELEMETRY_OPTOUT=1
 
 PATH=$PATH:~/.cargo/bin
 
-# Jump to path aliases
+# Jump to path and synchronization aliases
 __base_school='~/Work/School/FERI-RIT'
+
 alias jL='cd ~/Work/Linux'
 alias jS="cd $__base_school"
 alias jST="cd $__base_school/Tasks/MAG-2-1"
@@ -99,5 +100,7 @@ alias jA3M='cd ~/Work/Arma\ 3/Mods'
 alias jA3T='cd ~/Work/Arma\ 3/Tools'
 alias jIDI='cd ~/Work/IDI'
 
-# Sync aliases
-alias sS="rsync -azP --delete loki:$__base_school/Data loki:$__base_school/Theses $__base_school/"
+if [ $(hostname) = "odin" ]; then
+    alias pullS="rsync -azP --delete loki:$__base_school/Data loki:$__base_school/Theses $__base_school/"
+    alias pushS="rsync -azP --delete $__base_school/Data $__base_school/Theses loki:$__base_school/"
+fi
